+++ src/funding/funding.js
@@ -15,11 +15,6 @@ import { SUPPORTED_FUNDING_SOURCES } from "@paypal/funding-components/src";
 
 import type { Wallet, Experiment } from "../types";
 import { BUTTON_LAYOUT, BUTTON_FLOW } from "../constants";
-import type {
-  OnShippingChange,
-  OnShippingAddressChange,
-  OnShippingOptionsChange,
-} from "../ui/buttons/props";
 
 import { getFundingConfig } from "./config";
 
@@ -124,6 +119,7 @@ export function isFundingEligible(
     enableFunding,
     components,
     experiment,
+    flow,
     fundingSource,
     fundingEligibility,
     layout,
@@ -270,20 +266,16 @@ export function determineEligibleFunding({
 
 export function isWalletFundingEligible({
   wallet,
-  onShippingChange,
-  onShippingAddressChange,
-  onShippingOptionsChange,
+  hasShippingCallback,
 }: {|
   wallet: ?Wallet,
-  onShippingChange: ?OnShippingChange,
-  onShippingAddressChange: ?OnShippingAddressChange,
-  onShippingOptionsChange: ?OnShippingOptionsChange,
+  hasShippingCallback: boolean,
 |}): boolean {
   if (!wallet) {
     return false;
   }
 
-  if (onShippingChange || onShippingAddressChange || onShippingOptionsChange) {
+  if (hasShippingCallback) {
     return false;
   }