+++ src/funding/venmo/config.jsx
@@ -8,7 +8,7 @@ import {
 } from "@paypal/sdk-logos/src";
 import { DISPLAY_ONLY_VALUES, PLATFORM } from "@paypal/sdk-constants/src";
 
-import { BUTTON_COLOR, BUTTON_LAYOUT } from "../../constants";
+import { BUTTON_COLOR, BUTTON_LAYOUT, BUTTON_FLOW } from "../../constants";
 import { DEFAULT_FUNDING_CONFIG, type FundingSourceConfig } from "../common";
 
 import { WalletLabel, Label } from "./template";
@@ -17,9 +17,11 @@ export function getVenmoConfig(): FundingSourceConfig {
   return {
     ...DEFAULT_FUNDING_CONFIG,
 
+    flows: [BUTTON_FLOW.PURCHASE, BUTTON_FLOW.VAULT_WITHOUT_PURCHASE],
+
     layouts: [BUTTON_LAYOUT.HORIZONTAL, BUTTON_LAYOUT.VERTICAL],
 
-    eligible: ({ experiment, shippingChange, displayOnly }) => {
+    eligible: ({ experiment, shippingChange, displayOnly, flow }) => {
       /**
        * This could probably be removed if the enableVenmo experiment is
        * rolled out to 100%.
@@ -46,6 +48,13 @@ export function getVenmoConfig(): FundingSourceConfig {
         return false;
       }
 
+      if (
+        flow === BUTTON_FLOW.VAULT_WITHOUT_PURCHASE &&
+        experiment?.venmoVaultWithoutPurchase !== true
+      ) {
+        return false;
+      }
+
       return true;
     },
 
@@ -61,7 +70,9 @@ export function getVenmoConfig(): FundingSourceConfig {
         };
       }
 
-      return {};
+      return {
+        popup: true,
+      };
     },
 
     Logo: ({ logoColor, optional }) => {